pyproject.toml
[project]
name = "mcp-scripts"
version = "0.1.0"
description = "List of scripts for MCP"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.4",
]
